+++ src/cljs/rems/administration/create_form.cljs
@@ -513,7 +513,7 @@
                                          {:value :table :label (text :t.create-form/type-table)}
                                          {:value :date :label (text :t.create-form/type-date)}
                                          {:value :email :label (text :t.create-form/type-email)}
-                                         {:value :phonenumber :label (text :t.create-form/type-phone-number)}
+                                         {:value :phone-number :label (text :t.create-form/type-phone-number)}
                                          {:value :attachment :label (text :t.create-form/type-attachment)}
                                          {:value :label :label (text :t.create-form/type-label)}
                                          {:value :header :label (text :t.create-form/type-header)}]}])